# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api.NamespaceSyncStatus:
    properties:
      error:
        type: string
      flux_commit:
        type: string
      flux_ready:
        description: Backward compatibility fields (Phase 9)
        type: boolean
      git_commit:
        type: string
      last_applied_commit:
        type: string
      last_sync:
        type: string
      last_sync_time:
        type: string
      message:
        type: string
      namespace:
        description: New fields for Phase 17 frontend compatibility
        type: string
      secrets:
        items:
          $ref: '#/definitions/internal_api.SecretSyncInfo'
        type: array
      synced:
        type: boolean
    type: object
  internal_api.SecretSyncInfo:
    properties:
      commit_sha:
        type: string
      name:
        type: string
      status:
        type: string
      synced_to_k8s:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/yourorg/secret-manager/issues
  description: Production-ready secret management with drift detection and GitOps
    integration
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/yourorg/secret-manager
  title: Secret Manager API
  version: "1.0"
paths:
...
  /namespaces/{namespace}/sync-status:
    get:
      consumes:
      - application/json
      description: Get GitOps synchronization status between Git and Kubernetes via
        FluxCD
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sync status details
          schema:
            $ref: '#/definitions/internal_api.NamespaceSyncStatus'
        "400":
          description: Invalid namespace ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get FluxCD sync status
      tags:
      - sync
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"